Improvements based on comments on PR
